Product

+++ app/Traits/HandleImageTrait.php
@@ -1,36 +1,48 @@
 <?php 
 
 namespace App\Traits;
+use Intervention\Image\ImageManager;
+use Intervention\Image\Drivers\Gd\Driver;
 
-use Image;
+trait HandleImageTrait 
+{
+    protected $path = 'upload/';
 
-trait HandleImageTrait{
-    protected $path = 'uploads/';
-
-    public function veryfy($request){
+    public function veryfy ($request)
+    {
         return $request->has('image');
     }
 
-    public function saveImage($request){
-        if($this->veryfy($request)){
+    public function saveImage($request)
+    {
+        if($this->veryfy($request))
+        {
             $image = $request->file('image');
+            $manager = new ImageManager(new Driver());
+            
             $name = time() . '.' . $image->getClientOriginalExtension();
-            Image::make($image)->resize(300, 300)->save($this->path . $name);
+            $img =  $manager->read($image);
+            $img_s = $img->resize(300,300)->save($this->path . $name);
             return $name;
         }
     }
 
-    public function updateImage($request, $currentImage){
-        if($this->veryfy($request)){
+    public function updateImage($request, $currentImage)
+    {
+        if($this->veryfy($request))
+        {
             $this->deleteImage($currentImage);
+
             return $this->saveImage($request);
         }
 
         return $currentImage;
     }
 
-    public function deleteImage($imageName){
-        if(file_exists($this->path .$imageName)){
+    public function deleteImage($imageName)
+    {
+        if($imageName && file_exists($this->path .$imageName))
+        {
             unlink($this->path . $imageName);
         }
     }
